% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebook_browse}
\alias{notebook_browse}
\alias{notebook_peek}
\title{Preview the current notebook}
\usage{
notebook_browse(file = "notebook/docs/notebook.html")

notebook_peek(file = "notebook/docs/notebook.html")
}
\arguments{
\item{file}{path to the current notebook file. Defaults to `"notebook/docs/notebook.html"`}
}
\value{
`NULL`. The function is called for its effects on the project files.
}
\description{
`notebook_browse()` opens the current notebook in a browser.
`notebook_peek()` copies the notebook to a temporary file and opens it in the
RStudio Viewer pane.
}
