% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebook_reset}
\alias{notebook_reset}
\title{Reset the notebook directory to the last git commit}
\usage{
notebook_reset(notebook_path = "notebook", force = FALSE)
}
\arguments{
\item{notebook_path}{notebook folder. Defaults to `"notebook"`.}

\item{force}{whether to discard uncommited .rmd files. Defaults to `FALSE`.}
}
\value{
`NULL`. The function is called for its effects on the project files.
}
\description{
This function checks out the last git commit for the notebook folder. By
default, it throws an error if there are uncommitted .rmd files. This is a
dangerous function. Proceed with caution.
}
