% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebook_rmd_list}
\alias{notebook_rmd_list}
\alias{notebook_rmd_collate}
\title{Gather the paths to the notebook's rmd files}
\usage{
notebook_rmd_list(notebook_path = "notebook")

notebook_rmd_collate(notebook_path = "notebook")
}
\arguments{
\item{notebook_path}{notebook folder. Defaults to `"notebook"`.}
}
\value{
a vector of the paths to the rmd files.
}
\description{
`notebook_rmd_list()` lists the files in lexicographic order.
`notebook_rmd_collate()` puts them in knitting order (with the `index.Rmd`
first and the posts in reverse-chronological order).
}
