% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usedrake-package.R
\docType{package}
\name{usedrake-package}
\alias{usedrake}
\alias{usedrake-package}
\title{usedrake: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjmahr/usedrake}
  \item Report bugs at \url{https://github.com/tjmahr/usedrake/issues}
}

}
\keyword{internal}
